Client: Move websocket initialization from main app component to custom hook used by the component; move hooks file to be directly under src directory
