+++ client/src/components/conversation/conversation.component.tsx
@@ -29,7 +29,7 @@ import CustomAvatarGroup, {
 } from './custom-avatar-group.component';
 import ForwardMessageDialog from './forward-message-dialog.component';
 import MessageSeenByUsersDialog from './message-seen-by-dialog.component';
-import { useUserInfoData, useLazyLoading } from '../../pages/hooks';
+import { useUserInfoData, useLazyLoading } from '../../hooks';
 
 import {
   selectJoinedConversations,
