+++ client/src/pages/my-profile/my-profile-page.component.tsx
@@ -9,7 +9,7 @@ import { Box } from '@mui/material';
 
 import { AppState } from '../../redux/root-reducer';
 
-import { useLazyLoading } from '../hooks';
+import { useLazyLoading } from '../../hooks';
 
 import { User } from '../../redux/user/user.types';
 import { selectCurrentUser } from '../../redux/user/user.selectors';
